offst-bin: stnode dir refactoring.
