+++ components/bin/src/stnode/mod.rs
@@ -1,0 +1,6 @@
+mod file_trusted_apps;
+mod node;
+mod stnodelib;
+
+pub use self::node::{net_node, NetNodeError};
+pub use self::stnodelib::{stnode, NodeBinError, StNodeCmd};
