+++ components/bin/src/stnode/node.rs
